requirements.txt
kagglehub
pandas
numpy
matplotlib
seaborn
xgboost
scikit-learn
ipykernel
google-genai
python-dotenv